+++ src/main/java/kg/musabaev/megalabnews/service/impl/SimplePostService.java
@@ -36,6 +36,8 @@ import java.util.Set;
 
 import static kg.musabaev.megalabnews.service.impl.SimpleUserService.USER_CREATED_POSTS_CACHE_NAME;
 import static kg.musabaev.megalabnews.service.impl.SimpleUserService.USER_FAVOURITE_POSTS_CACHE_NAME;
+import static kg.musabaev.megalabnews.util.Utils.getLastPathSegmentOrNull;
+import static kg.musabaev.megalabnews.util.Utils.isAuthenticatedUser;
 
 @Service
 @Log4j2
@@ -104,8 +106,10 @@ public class SimplePostService implements PostService {
 			@CacheEvict(cacheNames = USER_FAVOURITE_POSTS_CACHE_NAME, allEntries = true)})
 	public void deleteById(Long postId) {
 		Utils.assertPostExistsByIdOrElseThrow(postId);
+		if (!isAuthenticatedUser(postRepo.findAuthorUsernameByPostId(postId)))
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
 		deleteImageInStorageIfExists(
-				Utils.getLastPathSegmentOrNull(postRepo.findPostImageUrlByPostId(postId)));
+				getLastPathSegmentOrNull(postRepo.findPostImageUrlByPostId(postId)));
 		Utils.deleteCommentsRecursively(postId, commentRepo.getAllRootCommentId(postId));
 		userRepo.deletePostsFromUserFavouritePosts(postId);
 
@@ -122,13 +126,16 @@ public class SimplePostService implements PostService {
 			@CacheEvict(cacheNames = POST_IMAGE_CACHE_NAME, allEntries = true)})
 	public NewOrUpdatePostResponse update(Long postId, NewOrUpdatePostRequest dto) {
 		return postRepo.findById(postId).map(post -> {
-			String imageFilename = Utils.getLastPathSegmentOrNull(dto.imageUrl());
-			String postImageFilename = Utils.getLastPathSegmentOrNull(post.getImageUrl());
+			if (!isAuthenticatedUser(post.getAuthor().getUsername()))
+				throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
 
-			postMapper.update(dto, post);
 			// если пред. название файла не совпадает с текущим, то удаляем пред. файл
+			String imageFilename = getLastPathSegmentOrNull(dto.imageUrl());
+			String postImageFilename = getLastPathSegmentOrNull(post.getImageUrl());
 			if (postImageFilename != null && !imageFilename.equals(postImageFilename))
 				deleteImageInStorageIfExists(postImageFilename);
+
+			postMapper.update(dto, post);
 
 			return postMapper.toDto(postRepo.save(post));
 		}).orElseThrow(() -> {
