+++ src/main/java/kg/musabaev/megalabnews/service/impl/SimpleUserService.java
@@ -33,8 +33,7 @@ import org.springframework.web.server.ResponseStatusException;
 
 import java.nio.file.Path;
 
-import static kg.musabaev.megalabnews.util.Utils.assertPostExistsByIdOrElseThrow;
-import static kg.musabaev.megalabnews.util.Utils.assertUserExistsByIdOrElseThrow;
+import static kg.musabaev.megalabnews.util.Utils.*;
 
 @Service
 @RequiredArgsConstructor
@@ -76,6 +75,8 @@ public class SimpleUserService implements UserService {
 	public void addToFavouritePosts(Long userId, AddToFavouritePostsRequest dto) {
 		assertUserExistsByIdOrElseThrow(userId);
 		assertPostExistsByIdOrElseThrow(dto.postId());
+		if (isAuthenticatedUser(userRepo.findUsernameByUserId(userId)))
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
 
 		try {
 			userRepo.insertIntoFavouritePosts(userId, dto.postId());
@@ -88,6 +89,8 @@ public class SimpleUserService implements UserService {
 	@Transactional
 	public void deleteFromFavouritePosts(Long userId, Long postId) {
 		assertUserExistsByIdOrElseThrow(userId);
+		if (isAuthenticatedUser(userRepo.findUsernameByUserId(userId)))
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
 
 		userRepo.deleteFromFavouritePosts(userId, postId);
 	}
@@ -106,6 +109,8 @@ public class SimpleUserService implements UserService {
 	@Cacheable(cacheNames = USER_CREATED_POSTS_CACHE_NAME, keyGenerator = "pairCacheKeyGenerator")
 	public Page<PostListView> getAllCreatedPostsByUserId(Long userId, Pageable pageable) {
 		assertUserExistsByIdOrElseThrow(userId);
+		if (isAuthenticatedUser(userRepo.findUsernameByUserId(userId)))
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
 
 		return postRepo.findAllByAuthorId(userId, pageable);
 	}
@@ -115,6 +120,8 @@ public class SimpleUserService implements UserService {
 	@CacheEvict(cacheNames = USER_ITEM_CACHE_NAME, key = "#userId")
 	public UpdateUserResponse update(Long userId, UpdateUserRequest dto) {
 		if (userRepo.existsByUsername(dto.username())) throw new ResponseStatusException(HttpStatus.CONFLICT);
+		if (isAuthenticatedUser(userRepo.findUsernameByUserId(userId)))
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
 
 		return userRepo.findById(userId).map(user -> {
 			userMapper.update(dto, user);
@@ -149,6 +156,8 @@ public class SimpleUserService implements UserService {
 			@CacheEvict(cacheNames = USER_PICTURE_CACHE_NAME, allEntries = true)})
 	public void deleteById(Long userId) {
 		assertUserExistsByIdOrElseThrow(userId);
+		if (isAuthenticatedUser(userRepo.findUsernameByUserId(userId)))
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
 
 		postRepo.findAllPostsIdByAuthorId(userId).forEach(postRepo::deleteById);
 		deleteImageInStorageIfExists(Utils.getLastPathSegmentOrNull(userRepo.findUserPictureByUserId(userId)));
